tests/wpt/web-platform-tests/web-bundle/subresource-loading/script-nested-bundle.https.tentative.html

<!DOCTYPE html>
<title>A nested bundle is not supported</title>
<link rel="help" href="https://github.com/WICG/webpackage/blob/main/explainers/subresource-loading.md" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/test-helpers.js"></script>

<script>
  window.TEST_WEB_BUNDLE_ELEMENT_TYPE = 'script';
  setup(() => {
    assert_true(HTMLScriptElement.supports('webbundle'));
  });
</script>
<body>
  <script type="webbundle">
    {
      "source": "/web-bundle/resources/wbn/nested-main.wbn",
      "resources": ["/web-bundle/resources/wbn/nested-sub.wbn"]
    }
  </script>
  <script src="resources/nested-bundle-test.js"></script>
</body>
